requirements.txt
asgiref==3.7.2
attrs==22.1.0
build==0.10.0
CacheControl==0.12.14
cachetools==5.3.1
certifi==2022.9.24
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
colorama==0.4.5
crashtest==0.4.1
distlib==0.3.6
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
dulwich==0.21.5
exceptiongroup==1.0.0
filelock==3.12.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
iniconfig==1.1.1
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.1
marshmallow==3.19.0
more-itertools==9.1.0
msgpack==1.0.5
mypy==1.1.1
mypy-extensions==1.0.0
numpy==1.23.3
oauthlib==3.2.2
packaging==21.3
pandas==1.5.0
pexpect==4.8.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.8.0
pluggy==1.0.0
poetry-core==1.6.1
poetry-plugin-export==1.4.0
protobuf==4.23.3
psycopg2==2.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.4
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==2.15.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.0.post1
six==1.16.0
SQLAlchemy==2.0.2
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typing_extensions==4.4.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.23.1
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.15.0
